requirements.txt
nltk==3.4.5
textstat==0.5.6
google-cloud-texttospeech==0.5.0
pydub==0.23.1
